# ENH: stats.pearsonr: two simple (but substantial) efficiency improvements